Refactor directory for auth, rol and user.
- Moved directories auth, user and rol into the new package called
  entity
- Updated imports and references across affected classes.
