+++ demo/src/main/java/com/project/demo/entity/rol/RoleEnum.java
@@ -1,4 +1,4 @@
-package com.project.demo.logic.entity.rol;
+package com.project.demo.entity.rol;
 
 public enum RoleEnum {
     USER,
