+++ demo/src/main/java/com/project/demo/entity/rol/RoleRepository.java
@@ -1,4 +1,4 @@
-package com.project.demo.logic.entity.rol;
+package com.project.demo.entity.rol;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
