+++ demo/src/main/java/com/project/demo/entity/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.project.demo.logic.entity.user;
+package com.project.demo.entity.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
