+++ demo/src/main/java/com/project/demo/rest/admin/AdminController.java
@@ -1,10 +1,10 @@
 package com.project.demo.rest.admin;
 
-import com.project.demo.logic.entity.rol.Role;
-import com.project.demo.logic.entity.rol.RoleEnum;
-import com.project.demo.logic.entity.rol.RoleRepository;
-import com.project.demo.logic.entity.user.User;
-import com.project.demo.logic.entity.user.UserRepository;
+import com.project.demo.entity.rol.Role;
+import com.project.demo.entity.rol.RoleEnum;
+import com.project.demo.entity.rol.RoleRepository;
+import com.project.demo.entity.user.User;
+import com.project.demo.entity.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.crypto.password.PasswordEncoder;
