+++ demo/src/main/java/com/project/demo/rest/auth/AuthRestController.java
@@ -1,13 +1,13 @@
 package com.project.demo.rest.auth;
 
-import com.project.demo.logic.entity.auth.AuthenticationService;
-import com.project.demo.logic.entity.auth.JwtService;
-import com.project.demo.logic.entity.rol.Role;
-import com.project.demo.logic.entity.rol.RoleEnum;
-import com.project.demo.logic.entity.rol.RoleRepository;
-import com.project.demo.logic.entity.user.LoginResponse;
-import com.project.demo.logic.entity.user.User;
-import com.project.demo.logic.entity.user.UserRepository;
+import com.project.demo.entity.auth.AuthenticationService;
+import com.project.demo.entity.auth.JwtService;
+import com.project.demo.entity.rol.Role;
+import com.project.demo.entity.rol.RoleEnum;
+import com.project.demo.entity.rol.RoleRepository;
+import com.project.demo.entity.user.LoginResponse;
+import com.project.demo.entity.user.User;
+import com.project.demo.entity.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
